requirements.txt
datasets==2.17.0
fqdn==1.5.1
isoduration==20.11.0
jsonpointer==2.4
jupyter==1.0.0
librosa==0.10.1
pip-chill==1.0.3
pydub==0.25.1
sentencepiece==0.1.99
torchaudio==2.2.0
transformers==4.37.2
uri-template==1.3.0
webcolors==1.13